refactor: extract utils and helpers from test file
